eXplainable-AI/requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1

Pillow~=8.2.0
requests~=2.25.1
tensorflow==2.4.0
shap~=0.39.0
matplotlib~=3.4.1
numpy~=1.19.5
azure-storage-blob~=12.8
onnx~=1.9.0
onnx2keras~=0.0.24